All tests pass.
Successfully worked out how to set the root_path to go to "sessions#new" so hitting '/' goes straight to the login page (and negating the need for after_sign_out_path_for) but then realised I don't actually want to do that because hitting the homepage when logged in would go there too.

Tried to set user_root so that after login users would go straight to users#show without the hacky after_sign_in_path_for override but Devise's after_sign_out_path_for assumes that the user_root doesn't need a resource and calls it without one. So either way the method would need to be overridden.

Probably want to create a home_controller#index action that says 'if logged in send to users#show else send to login' and maintains flash messages properly and leave user_root as is.